Dev to main

+++ src/controllers/user/user.js
@@ -1,24 +1,24 @@
 const { User} = require('../../db');
 const { v4: uuidv4 } = require('uuid');
-const { Op } = require("sequelize");
 
-const exclude = ['createdAt', 'updatedAt']
 
 async function newUser(req, res, next) {
-    if (!req.body.displayName || !req.body.email || !req.body.password) {
-        return res.status(400).json({ message: 'Bad request' })        
+
+    if (!req.body.name || !req.body.email || !req.body.password) {
+        return res.status(400).json({ message: 'Bad request' })
     }
-    const photoURL=null;
+    const photoURL="https://i.imgur.com/vfrW9Xx.png";
     if(req.body.photoURL)photoURL=req.body.photoURL
-    const { email, displayName, password} = req.body
+    const { email, name, password} = req.body
+    const  id=uuidv4();
+    let user={id,email,name,password,admin:false,photoURL};
     try {
-        const exist = await User.findOne({ where: { email: email } })
-        if (exist !== null) { return res.status(500).send({ message: 'El email ya existe.' }) }
-        const exist2 = await User.findOne({ where: { displayName: displayName } })
+        const exist = await User.findOne({where:{email:user.email}})
+        if (exist) { return res.status(500).send({ message: 'El email ya existe.' }) }
+        const exist2 = await User.findOne({ where: { name: user.name } })
         if (exist2 !== null) { return res.status(500).json({ message: 'El nombre de usuario ya existe.' }) }
         const id = uuidv4()
-        const user = { id, displayName, password, email, photoURL }
-        
+
         await User.create(user)
         return res.send(user)
     } catch (error) {
@@ -30,7 +30,7 @@ async function updateUser(req, res, next) {
     const { idUser} = req.body
     try {
         const user = await User.findByPk(idUser)
-        req.body.displayName ? user.displayName = req.body.displayName : '';
+        req.body.name ? user.name = req.body.name : '';
         req.body.password ? user.password=req.body.password:'';
         req.body.photoURL?user.photoURL=req.body.photoURL:'';
         user.save()
@@ -42,23 +42,11 @@ async function updateUser(req, res, next) {
 
 
 async function getAllUsers(req, res, next) {
-    let {displayName, admin} = req.query
-    if(displayName === 'undefined') displayName = ''
+    let {name, admin} = req.query
+    if(name === 'undefined') name = ''
     if(admin === 'undefined') admin = undefined
     try {
-        const user = await User.findAll({
-            where: admin !== undefined ?
-            {
-                displayName: { [Op.iLike]: `%${displayName}%` },
-                admin
-            } : {
-                displayName: { [Op.iLike]: `%${displayName}%` }
-            },
-            attributes: {
-                exclude: [...exclude,'password']
-            },
-            order: ['displayName']
-        });
+        const user = await User.findAll();
         return res.send(user)
     } catch (error) {
         next({ message: 'Bad Request' })
@@ -71,6 +59,8 @@ async function deleteUser(req, res, next) {
     }
     const { idUser } = req.params;
     try {
+        const local=await User.findByPk(idUser);
+        if(!local) return res.send('El usuario no existe.')
         await User.destroy({
             where: {
                 id: idUser
@@ -85,8 +75,8 @@ async function deleteUser(req, res, next) {
 
 
 async function loginUser(req, res, next) {
-    const {email, displayName, password} = req.body
-    if(displayName) {
+    const {email, name, password} = req.body
+    if(name) {
         try {
             const isUser = await User.findOne({
                 where: {
